servo: ping device to check power state
The servo charger code was checking the power status of the system by
querrying the state from the EC. The EC reporting an AP state of S0 does
not mean that the system can be communicated with though. Instead ping
the system to check if it's up.

BUG=b:168939843
TEST=test_that --board <board> <IP of DUT> power_WakeSources

Change-Id: I258c70304750dd7d6809bcfcf3a824a74534fc30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2485881
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>