bluetooth: Refactor BluetoothXmlRpcDelegate
Rather than start a separate xmlrpc server for Bluetooth, we will re-use
the multimedia xmlrpc server and expose Bluetooth as a native facade.
This allows us to re-use some of the retry and recovery mechanisms built
into the multimedia facade factory.

The original BluetoothXmlRpcDelegate is kept around and it inherits from
the BluetoothFacadeNative now. Eventually we will want to remove this
usage and replace it with calls to the multimedia xmlrpc server.

BUG=b:169041126
TEST=Ran bluetooth_AdapterQuickHealth

Change-Id: Ida325b23ed09d082daeb4a7d2785f242341134ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2489242
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>