bluetooth: Catch all errors on test wrapper
Anytime there is an error raised inside a test wrapper, report it as
a failure and continue running the rest of the batch. This makes it
possible for the batch summary to be generated even when an exception is
raised in one fo the tests.

BUG=b:154564063
TEST=Raise random exception in a test and make sure the rest of the
     batch is run

Change-Id: Ifa32b29eecb9c6e71ef1768812022b8640e7c0fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2500406
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Yu Liu <yudiliu@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>