power_WakeSources: simplify style
The test used both "str.startswith" and checking membership in a list to
check if a wake source was one of 2 wake sources. This changes both to
the list membership check.

It also changes some if statements that would all return True to elif
statements with a return True at the end to maximize code visibility.

BUG=None
TEST=test_that --board <board> <DUT ip> power_WakeSources

Change-Id: I3692d42f8028daa3229693b4bed450c1f5c1a749
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2506550
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>