power_WakeSources: Test wake on AC when supported
Tests that wake on AC connect/disconnect causes a dark resume if the
wake mask (as reported by the EC) indicates the feature is enabled for
S0ix/S3. It also refactors the test to remove duplicate code around
_test_wake, _is_valid_wake_source, etc. and uses the power button
instead of the RTC as the backup wake source.

BUG=b:168939843
TEST=test_that --board <board> <IP of DUT> power_WakeSources

Change-Id: Ic0f59cc04a4efaf990af45d01e3a304b9f378e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2477339
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>