firmware_test: remove the '&' from the command
In order to use the Popen.communicate, we can't add the
'&' to the command to force it running in background
because the Popen.communicate will be blocked until the
background process exit. On the other hand, we can pass
the block=false argument to the run_shell_command to force
the process runs as unblocking.

BUG=b:171771020, b:172325331
TEST=No hang when running test_that -b ${BOARD} with below tests
     1. firmware_ECWakeSource
     2. platform_S0ixCycle
     3. platform_S3Cycle
     4. firmware_ECUsbPorts
     5. firmware_Cr50DeviceState

Change-Id: Idd0e3395bfe43b638618be07a9215849f01cdbca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2505232
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>