bluetooth: WBS tests with cras default codec selection
If a board supports WBS, the default selection in cras
should be WBS. On the other hand, if a board does not
support WBS, the default selection should be NBS.
Conduct the WBS/NBS tests with their default WBS/NBS
selection. Disable WBS only when a board does support
WBS and a NBS test is to be conducted.

BUG=b:171928376
TEST=Conduct bluetooth_AdapterAUHealth tests with
the test image R88-13543.0.0:
- On an octopus: all audio tests should pass.
- On a hatch: the WBS tests should fail due to a regression
              about selecting the default codec. Other audio
              tests should pass.

Change-Id: I86b6658c1e32c0ac2d32e78f4769e4e02fa4f5d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2506973
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>