autoupdate_Basic: Check for correct kernel before verifying logs.
Seeing this error: "list index out of range" when trying to verify the
hostlog of an update that just completed. It is trying to access the
update_engine.log for the update that just happened before reboot but
they are not available.

So check that we are in the right partition before verifying the logs.
Also make the error message clearer when this happens.

BUG=chromium:1143742
TEST=autoupdate_Basic

Change-Id: I211e9cd076c8bbdcae01b03f9513d4dbb81defdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2508455
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>