network_WiFi_APSupportedRates: skip for Scarlet
We're not backporting this to 4.4.

Ideally, we'll port this to Tast and get a proper software dependency
(so this only skips kernel 4.4 + ath10k, and we can pick up the test
when RK3399 platforms move to newer kernels). Use this as a hack for
now.

BUG=b:138406224
TEST=run it on Scarlet and other BOARDs

Change-Id: If5768280e44bee3106caa503481bcef2e9f439fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2530269
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Claire Chang <tientzu@chromium.org>
Reviewed-by: Claire Chang <tientzu@chromium.org>