firmware tests: Limit number of retries
Now that tests are less flaky we can limit the number of retries:
If a test fails, it will probably fail all the time (and take a long
time for each of them, waiting for whatever error condition pops up),
giving us a very skewed runtime: either a test completes quickly
(success) or it takes ~forever on failure.

JOB_RETRIES defaults to 1, and "no retries" has to be specified
explicitly. Therefore, this change drops JOB_RETRIES on all firmware
tests except if it's configured to 0 already - a property we want to
keep and even extend to more tests, confidence permitting.

BUG=b:171297016
TEST=TODO: run faft_bios and see how bad it is

Change-Id: I30c398d60b6376ddc253d685b399eef2cff50dcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2532220
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>