network_WiFi_VerifyRouter: Support pcap device verification.
In b/172221785, we noticed a pcap device that cannot communicate with
DUT and router. In some tests, we rely on a health pcap to provide
packet capture result to verify the behavior. It would be helpful if
VerifyRouter test can also sanitize pcap device.

Also, we started to use pcap as router in some test, so verify pcap
can be more important if we add more tests using it.

Extend the test to support verifying pcap device and add a separate
control file for it.

BUG=b:172221785
TEST=test_that network_WiFi_VerifyRouter
TEST=test_that network_WiFi_VerifyRouter.pcap

Change-Id: I9fc5e6224c8077ee1125b94cea0c2d75746b919f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2534410
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>