mode_switcher: Change the sequence of booting to rec
Switch the USB key when AP is _on_, because there is a bug
(b/172909077) - using "image_usbkey_direction:usb_state"
(using image_usbkey_mux insead works - only switch the muxer, don't
power off/on the USB stick during that process), when AP is off may
cause not recognizing the file system, so system won't boot in recovery
from USB.

At the moment, the issue is not root-caused. Before that, apply
the workaround to make the faft suites more stable.
Once the bug is fixed, changing the muxer can be done when AP is off.

BUG=b:172909077, b:170894065
TEST=Run firmware_EventLog on Scarlet

Change-Id: Ief4d3c579822a7d68a3e9f88fdd3797e9ac86e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2536912
Tested-by: Dawid Niedźwiecki <dn@semihalf.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>