bluetooth: Make TestNA override failures
If we see a TestNA emitted in code, use it to override any failures seen
in the same test run. As a result of this change, all TestNA errors in
code mean that the entire test suite result is Not Applicable anymore
instead of any one statement/expression.

BUG=b:173430197
TEST=Forced discovery to always return False and verified tests failed
     with TestNA.

Change-Id: Ia87b392ca3debf8e6aa516f316ce0af1879c180e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2543463
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>