firmware_WriteProtectFunc: decouple from chromeos-firmwareupdate
The test used the image shipped in the firmware updater which
creates issues if that image and the written image differ in certain
significant ways (e.g. the fmap layout). Rework it to use the flashed
firmware and go from there.

BUG=b:172330099
TEST=test_that -b eve 10.99.0.100 --fast --no-retries
firmware_WriteProtectFunc passes twice in a row

Change-Id: I5db5950cf4efcb1f4419532b3eda3e6c1e2aed9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2545832
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>