servo_charger: increase Timeout for ssh communication
When running a role change on v4, the usb device(s) reenumerate on the
DUT. This takes a few seconds. Then the DUT needs to reconnect to the
network in cases where the network connection was coming from the v4's
ethernet adapter.
This change extends the timeout to 1 minute to allow comfortably for
this check over ssh to work.

BUG=None
TEST=test_that --autotest_dir . $DUT_IP power_ServoChargeStress.3loops
/tmp/test_that_results_eM1TYR/results-1-power_ServoChargeStress.3loops
[  PASSED  ]

Change-Id: I6c180930486851f3b1067ca5791f3136023e9346
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1745894
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>