autotest: Split device health profile to load separate.
This CL is preparation to the next to init device-profile during creating servo-host.
The split will allowed to init device-profile as soon we confirmed that we have ssh access to the labstation and use it for servo-verify/repairs.

BUG=b:163845694
TEST=run local

admin-repair (servo_v4)
./server/autoserv -s --host-info-subdir host_info_store -m chromeos1-row4-rack4-host3 --lab True --local-only-host-info True -R -r /tr/

admin-repair (servo_v3)
./server/autoserv -s --host-info-subdir host_info_store -m chromeos1-row4-rack1-host5 --lab True --local-only-host-info True -R -r /tr/

Change-Id: I2b48ed2e47e5b836c2210ea6032916398d4d7900
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2554081
Tested-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Otabek Kasimov <otabek@google.com>