power_BatteryDesignCapacity: test to verify battery capacity
This test verifies that the battery full charge reflects the battery
design capacity after the device has been used for a few months.

If the battery full charge is usually lower than the battery design
capacity, then it is not fair to use the battery design capacity in
tests such as power_LoadTest. In this case, apply a discount on the
battery design capacity like CL:2417836 and CL:2521653.

BUG=b:162360640
TEST=run power_BatteryDesignCapacity

Cq-Depend: chromium:2705458
Change-Id: Id526f40413188b3eba6c478b08313e097c01eb8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2559597
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Commit-Queue: Mengqi Guo <mqg@chromium.org>
Tested-by: Mengqi Guo <mqg@chromium.org>