Check for post-update notification in autoupdate_UpdateFromUI
Add a check for the "Update available" notification that appears when
the device is ready to reboot after the update. The client test now
leaves the Chrome session logged in after it checks for the
notification, so the DUT won't be auto-rebooted like in the first
attempt crrev.com/c/2559602.

BUG=chromium:1152563
TEST=test_that localhost:2222 autoupdate_FromUI.full -b eve --no-retry --autotest_dir ~/trunk/src/third_party/autotest/files --args="job_repo_url=http://100.115.168.195:8082/static/eve-release/R89-13621.0.0/autotest/packages running_at_desk=True"

Change-Id: Ife2e63ba7639c5a5992d4023308d00aac3c8a3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2566879
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>