autotest: Stop powerd on labstation before reboot
That is work around to force labstation to boot when we provide power for
device. Some labstation is not booting by default possible because
device remember last power state of the device: if it was off then it
will stay off when we deliver the power to the device.

BUG=b:174514811
TEST=run local

Change-Id: If4644ec4461e26ab760cac67d96b7952e0373425
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2568909
Commit-Queue: Otabek Kasimov <otabek@google.com>
Tested-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>