[Autotest][PY3] Update client/bin/result_tools for python3 compatibility
Adjusting the imports in result_tools to support python3.
These ones are a bit weird, result_tools can be run **in** autotest, or as a
standalone lib. Because of this we need to adjust imports to be non-relative
for autotest support, but also leave the relative (under the except) for
standalone support

BUG=chromium:990593
TEST=dummy_pass in python 2 and 3. applicable unittests

Change-Id: I38d425c1d0a19aa7f7981e996b7b03a651699e72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2570083
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>