bluetooth: Add test coverage for EIR name >48 bytes
This adds a new test, sa_eir_max_name_size_test, that applies a bluez
alias longer than the default EIR size, and validates that the data is
not truncated when registered with the controller. The new test was
confirmed to pass when the EIR max size is defined, and fails when the
EIR max size is not defined (defaults to 48 bytes) on kukui.

BUG=None
TEST=SAHealth batch, new test with and without config definition

Change-Id: I0cba70b5e56d0a112f724c50bd0c5dbe23235a44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2575055
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>