config_unittest: Remove integration test
The removed test checks whether each JSON file in fw-testing-configs.py
has a correct-looking "platform" attribute. This is not a unit-test; it
is an integration test. It is redundant; we test that same information
in the fw-testing-configs repo. It has caused a CQ failure before in
Autotest. Let's remove that test from Tauto.

BUG=b:171419697
TEST=config_unittest.py

Change-Id: I4e07368eeac723aff79065551bbb53bc30b21a7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2577866
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>