Add faft_experimental to attribute allowlist
This suite was originally added by https://crrev.com/c/1925545. It
already has a test_suites/control file. However, it was not being used
for any tests, so it seems to have been removed from the
attribute_allowlist. The related CLs here use the suite, so we need to
add it back to attribute_allowlist.

BUG=b:175240134
TEST=None

Change-Id: I7b241e315e0719d28b4190d999de861807668e94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2582511
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>