[Autotest][PY3] cros/cryptohome rework proposal
tldr, crpytohome has 2 things going  for it:
1.) Functions which do things related to cryptohome, that don't need any
special libs (they run using utils.run).
2.) Some stuff with gobjects, dbus, mainloop, etc

The second item is not easy at all to move to py3. gobject is not in
py3, the replacement (    from gi.repository import GObject as gobject)
isn't in tauto deps. uprevving packages is not easy, but can be done.

So right now my proposal is to remove the functions out of cryptohome
that are completely stand alone, and use it instead, with the migration
of main cryptohome to happen later.

The utils are on the semi-critical path to tauto in py3, but the full
lib is not.

Open to suggestions!!!

Change-Id: I3e83342343141840d610e550bba4c8d40f117411
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2587471
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@google.com>