[Autotest][PY3] Add a sys.executable to the autoserv cmd
This will allow autoserv to be called with the version of python that is
being used on test_that. The basic idea is: we will set test_that into
the proper version of python, then all downstream scripts/args will be
called with a sys.executable in the cmd, so they are kicked off using
the same python runner as the launcher.

BUG=chromium:990593
TEST=dummy_Pass (python2 and 3), server and client

Change-Id: I54d6a13155ba1de42c0d9a96d28c58e66849fabc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2589563
Tested-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>