bluetooth: Use device time to make sure suspend is recent
If the time on the test runner is different from the DUT,
test_wait_for_resume might raise TestNA because the suspend attempt is
not seen as recent. To fix this, we send a test start timestamp to the
resume function that is taken from the DUT so that we can raise TestNA
only if the suspend occurred before the test started.

BUG=b:175639556
TEST=Ran bluetooth_AdapterSRHealth with correct DUT time and DUT time
     set to 1 hr behind workstation time

Change-Id: I37e62a5d065a63a92899c2844ff5c82cc3ef056a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2593733
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>