bluetooth: Skip RSSI filtering test if HW offloading not available
By default the LE_SCAN_FILTER_DUP flag is enabled on all platforms.
Due to this, the host does not receive as many advertisements during
passive scanning, which causes SW based RSSI filtering not to work
as intended. So, if the controller offloading is not supported, skip
the tests that involves RSSI filtering and raise TEST_NA.

BUG=b:175627184, b:169375729, b:173153478
TEST=1. Run bluetooth_AdapterAdvMonitor tests and verify single client,
        multi client, suspend/resume tests do not run and raise TestNA.

Change-Id: I8a70eabdbd1cb87069e0dd32e1e4fcbb9d635a78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2591822
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>
Commit-Queue: Manish Mandlik <mmandlik@chromium.org>
Tested-by: Manish Mandlik <mmandlik@chromium.org>