[Reland][Autotest][PY3] Update utils.py to support Python3
Relanding this: crrev.com/c/2593727

Having to do it manually because it won't let me reland...

Issue turned out to be a downstream test lib is checking for the
response to be type str, but in python2 the .decode() returns type
unicode (which is not a str in python2 >_<). So adjusting the decoding
to follow a bit of logic, which should be safer.

Change-Id: Ie3bd171b56944becc84aa77f94b5af7faa294140
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2593509
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>