Add control file and suite for fingerprint MCU
The fingerprint team has some tests that need to run on special setup,
which is a Chromebook attached with a standalone fingerprint MCU board.
Because these tests need to be scheduled on only those setups, we need
a new suite definition.

These files will be associated with a suite_scheduler config
(http://crrev/i/3462030) and a Tast attr.

BUG=b:158580909
TEST=none

Change-Id: I4efa2eb3837450d3fa96b24745ec554f9479cf49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2593741
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>