toolchain-tests: update the Makefile to always run compilation tests
We had to explicitly run `make check` here in order to get compile-time
checks, which we aren't currently doing. Since it's easy for these
diagnostic tests to accidentally break, and they have all of the same
dependencies as the test file's, fold these tests into the rules for
object file generation.

We could theoretically get slightly better parallelism by breaking these
out, but it takes <1 second for this to finish, so simplicity is
probably the thing to shoot for here.

BUG=chromium:1159199
TEST=emerge-amd64-generic chromeos-base/toolchain-tests with broken
     diagnostic tests.

Change-Id: I452a6bdf0859c653266e85d125dcaaa5cae24c6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2594294
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>