client/site_tests/firmware_LockedME: Removes ME region is readable check
The patch removes check which verifies ME region is readable or not when
the end of manufacturing is disabled. As CSE Lite SKU is integrated into
coreboot, Host needs to have read access to ME region to determine the
interrupted CSE FW update. Hence, during RO lock Host's master access
permissions are configured such that Host can read but not write to the
ME region.
The patch also appends 'SI_DESC' to the inaccessible_sections[] list so
that check_region_inaccessible() checks writability to all the regions.

BUG=b:175737945

Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.corp-partner.google.com>
Change-Id: I6fc03e11a97ba3102ba355d5c6bd7ca15d031396
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2594620
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>