drivefs-uprev: Add DriveFS CQ test suite
DriveFS is a package in ChromeOS that is built in google3 and uprevved
into ChromeOS using PUpr. To qualify a new uprev we need to ensure all
critical tast tests and also tests which have a software dependency on
DriveFS are ran and pass. This suite introduces the former.

BUG=chromium:1159653
TEST=None

Change-Id: Icb6d4606d7d19d9316b81d9de925a1b49d727393
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2595555
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@google.com>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>