Bluetooth: Modify peer reset to accept device_type as args
If we know what peripheral we will be emulating, send this information
as an argument to the peer reset call. This allows the peer to take
custom initialization steps depending on the desired device type.

This change is tested with the corresponding Chameleon change to ensure
that the optional arguments are sent over the xmlrpc server properly, as
some older versions of xmlrpc server do not allow optional args.

BUG=b:172381798,b:169633231
TEST=QuickHealth and other sub-tests on Barla and Kukui

Change-Id: I1e4b08d3213599502b6e57ccbe4d5f8bfc88e912
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2601296
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>