Accept 'Not charging' as charging when soc is nearly full
When a battery is nearly full, a gas gauge may trigger a full charge
protection and stops charging even with AC power is available.
If the charge current is 0, newer kernels present this state as
'Not charging'.

This patch makes firmware_ECChargingState accept this state (nearly
full thus not charging) as charging.

BUG=b:176992437
TEST=None

Change-Id: I8d4b4b33dabd51f590a1bed92aaad22c038b82c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2618493
Reviewed-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>