autotest: Add support for throughput checking to WiFi perf tests
This change adds support for individual test control files in the WiFi
performance suite to define a set of minimum throughput thresholds that
devices are required to meet. If a test run results in throughput values
below the defined threshold, the test will output a message indicating
the failure and continue running. After the entire test has been
completed, the test will fail if any throughput thresholds were missed
in any of the test runs.

The reasoning behind this approach is documented here:
http://doc/1qguLGgA_Xv4P_9anblRte8tF2xG549o2i4ia60g2QI4

BUG=b:177004163
TEST=Run the ht20 test an check that it fails and passes correctly based
on dummy throughput configs.

Change-Id: I30bc0868411c35d6210f14fc25b96027bb0fa683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2628178
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>