bluetooth: update peers in a small pool to next commit
A few DUTs in the lab will serve as a pre-verification pool
against the next chameleon bundle commit to verify.

In particular, the lab_next_commit will be used only when 3
conditions are satisfied
- the lab_next_commit is non-empty
- the hostname of the DUT can be found in lab_next_models
- the host_build of the DUT is the same as lab_next_build

If any of the above conditions is not satisfied, a DUT will
keep using the original stable lab_curr_commit.

BUG=b:177501379
TEST=Set the bluetooth_commits.yaml file and watch the dashboard.

Change-Id: Ide00f975a7ef0a988d2b970df655d88a88fa3400
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2629407
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>