Bluetooth: detect USB disconnects during tests
A common failure in the lab is for either the bluetooth controller to
detach from the USB bus or for some platform issue to cause the entire
usb bus to drop, disconnecting our usb device. In these cases, this
patch will identify the proper bluetooth usb device, watch for
disconnect events, and surface it up as a failure reason if it occurs.

BUG=b:173247857
TEST=Verified that usb source is identified and added to COMMON_FAILURES

Change-Id: Id6bfa6f5f0c6cb9f9b260eb607f49183cb58eb12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2670128
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>