[PY3][Autotest] First go at setting test_that flag for py3
NOTE: its gaurded with a force back to py2 for now. But this is roughly
what I think I am going for. There is a bit of debate I am having still
about turning on the "no restart" flag going forward. IMO the launching
script, (autoserv, test that, or other bins) should be able to control
the version they want to run in going forward. 1 pager about this likely
coming

BUG=chromium:990593
TEST=test_that with the py_version flag in both 2/3 and verfied tests
are in py2, as well as without the flag

Change-Id: If34d6866061f93e72728a3e49c02e2faf73ae628
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2638720
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>