mode_switcher: check all ports for USB stick
In the check_vbus_and_pd_stat function, check every port when verifying
if DUT is DFP (to see USB stick). There are cases when micro_servo is
connected and servo_v4 is plugged in port 1.

Also ensure that all checks are done when servo_v4 type-c is used,
because verifying pd state for type-a does not make sense.

BUG=b:173260514
TEST=run firmware_UserRequestRecovery with servo_micro and servo_v4
connected to the port 1 on a board with VBUS bug in RO e.g. coral

Change-Id: I971b8abd13deda3edfc132a82ff4bb9fb15c76ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2640416
Tested-by: Dawid Niedźwiecki <dn@semihalf.com>
Commit-Queue: Philip Chen <philipchen@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Philip Chen <philipchen@chromium.org>