hardware_StorageFio: Log blkdiscard
Log when we attempt the blkdiscard command. At the moment we only see if
the blkdiscard command fails in the logs, but we want to know if we attempted
at all. This is to allow us to confirm that we are doing the blkdiscard as expected.

BUG=b:178529109
TEST=ran hardware_StorageFio on hatch device, saw log

Change-Id: I58952548ddc13b922a76f8fb75449ab167981db4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2649157
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>