Bluetooth: Validate scan response in peer advertising tests
We learned in some cases the controller will accept a scan response and
not use it, so this tests will validate the scan response we registered,
if it exists, in all peer tests.

While we're at it, improve the failure cause reporting in peer tests.

BUG=b:178625794
TEST=AdvHealth on kukui and hatch with scan response enabled

Change-Id: I6d24aea595d3f936462450be048ccd2e812dd78a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2654198
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>