Revert frame rate change to graphics_HwOverlays.video
The frame rate was increased in this test to deal with failures in other
tests that were occurring on volteer due to it enabling overlay
prioritization. This actually caused the test to become flakey across
the board, so we are reverting this change. This may leave volteer
flakey, but we will determine that later. It always passed when I ran
it.

BUG=b:178660666
TEST=graphics_HwOverlays.video passes on volteer

Change-Id: I7ab1720a53e7a1580c0e917a85485d51f758a491
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2659656
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Auto-Submit: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>