[PY3][Autotest] Making Python 2 configparser for Py3.
The configparser, and six.configparser, in python3 do not work with our
.ini. Originally I planned just to change the .ini and call it a day,
but then it doesn't work in Python 2. This issue is made worse by ~20+
configs in the .ini would need to be adjusted to play nice in 2 & 3, and
the sources (lots, and scattered all over) updated too.

Instead, we will do some light parsing of the config file to handle the
subtle, but breaking differences between py2/3

BUG=chromium:990593
TEST=dummy_Pass, CQ, unittest in py2/3

Change-Id: I090917f513ba25ff396a9aa5936d6119fe3ea4cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2664729
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>