autotest: Add minimum throughput values for WiFi perf tests
Add values to the WiFi perf test control files to enforce minimum
throughputs. This will cause the tests to fail if the observed
throughput in test runs is below the provided values. The process for
deciding these values is described in this design document:
http://doc/1qguLGgA_Xv4P_9anblRte8tF2xG549o2i4ia60g2QI4

BUG=b:177004163
TEST=Run various perf tests and check results.

Change-Id: I3f1360b2f360666c49d77b71c02bc5bd01b1088d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2660137
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>