Add `tast_expr_b64` to allow users to pass base64 encoded tast expr
In the previous CL, I was trying to make this test user-configurable.
However it is hard for Skylab, due to crbug/1173329.
Tast expressions, like (“group:mainline” && “dep:lacros”) is a
toil to design escape patterns for single/double quote and
space. This CL adds a new flag, which applies base64 to encode and
decode the expressions without any restrictions.

BUG:1152839
TEST:test_that and skylab cmd(gpaste/6412444164423680)

Change-Id: Ia95a92ed043352039d259e52e8e38177f3628c6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2666384
Commit-Queue: Xinan Lin <linxinan@chromium.org>
Tested-by: Xinan Lin <linxinan@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>