cheets_GTS_R: Update the no-test module list.
I guess it is a error in test build config.
This is a new helper app added in ag/13192768,
which should have not been added as a test module.

Until this is fixed in upstream, we can ignore
these no-test moduls.

BUG=b:170182905
TEST=cheets_GTS_R.8.0_r3.GtsInstallerV2

Change-Id: I3be0f69567d8a569b592b5ab55ae1a71ad3d84fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2674688
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Jiyoun Ha <jiyounha@chromium.org>
Reviewed-by: Jiyoun Ha <jiyounha@chromium.org>