servo: Create servo.get_servo_v4_type()
Getting the Servo control 'servo_v4_type' directly is risky, as not all
servos have that control. This has already led to the attached bug.

The proper way to get the servo_v4_type is to first check whether the
servo_type is v4 (or, equivalently, whether the servo has the
servo_v4_type control), and *then* to get 'servo_v4_type'.

That logic has already been implemented in several places. This
repetition is prone to error, such as in the attached bug. It is safer
and cleaner to create an abstraction.

BUG=b:179686779
TEST=firmware_FAFTSetup & firmware_ECUsbPorts on servo, not v4
TEST=firmware_FAFTSetup & firmware_ECUsbPorts on servo v4, not type-c
TEST=firmware_FAFTSetup & firmware_ECUsbPorts on servo v4 type-c

Change-Id: I2c5c2f8d804d02f7e15526ebd4b12309c2e5688f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2681127
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Greg Edelston <gredelston@google.com>
Auto-Submit: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>